# TreckSec_API_Documentation.yaml
openapi: 3.0.0
info:
  version: '2.0'
  title: "TrekSec OpenAPI 3.0"
  description: API for a secure trekking.
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Localhost
paths:
  /users:
    post:
      description: >-
        Creates a new user in the system.
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: 'User created. Link in the Location header'
          headers:
            'Location':
              schema:
                type: string
                description: Link to the newly created user.
        '409':
          description: 'User already exists'
          content:
            application/json:
              schema:
                type: string
                description: User already exists error message.
        '400':
          description: 'Error message sent, with the actual error'
          content:
            application/json:
              schema:
                type: string
                description:  error message.
    get:
      description: >-
        Gets the list of users.
      summary: View all users
      responses:
        '200':
          description: 'Collection of users'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      description: >-
        Get one user by id.
      summary: View a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'data of a user'
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}/coordinates:
    get:
      description: >-
        Get the users coordinates by the id.
      summary: View a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'data of a user'
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User' 
    put:
      description: >-
        update the coordinates of the user
      summary: Update the coordinates
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - coordX
                - coordY
              properties:
                id:
                  type: integer
                  description: 'id of the user'
                coordX:
                  type: string
                  description: 'latitude od the dangers position'
                coordY:
                  type: string
                  description: 'longitude od the dangers position'
      responses:
        '200':
          description: 'Coordinates changed successfully'
          content:
            application/json:
              schema:
                type: string
                description:  Users email
        '404':
          description: 'Error message sent, with the actual error'
          content:
            application/json:
              schema:
                type: string
                description:  User not found
        '500':
          description: 'Si è verificato un errore durante il reset della password. Riprova più tardi'
          content:
            application/json:
              schema:
                type: string
                description:  Internal server error.
  /users/reset-password:
    post:
      description: >-
        gets the mail (used to send an email) to reset the password.
      summary: Send an email
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  description: 'Email address of the student'
      responses:
        '201':
          description: 'Email sent to the user that required the reset assword'
          content:
            application/json:
              schema:
                type: string
                description:  Mail sent message.
        '404':
          description: 'Email not found'
          content:
            application/json:
              schema:
                type: string
                description:  Email not found message.
              
    put:
      description: >-
        update the password of the user
      summary: Update the password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - newPassword
                - newcheckPassword
              properties:
                token:
                  type: string
                  description: 'Token of the user'
                newPassword:
                  type: string
                  description: 'newPassword of the user'
                newcheckPassword:
                  type: string
                  description: 'Newcheck of the users password'
      responses:
        '201':
          description: 'Password changed successfully'
          content:
            application/json:
              schema:
                type: string
                description:  Mail sent message.
        '400':
          description: 'Error message sent, with the actual error'
          content:
            application/json:
              schema:
                type: string
                description:  error message.
        '500':
          description: 'Si è verificato un errore durante il reset della password. Riprova più tardi'
          content:
            application/json:
              schema:
                type: string
                description:  Si è verificato un errore durante il reset della password. Riprova più tardi message.
...
  /users/telefoni/{telefono}:
    get:
      description: >-
        Get one user by phone number.
      summary: View a user
      parameters:
        - in: path
          name: telefono
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'data of a user'
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/User'
        '404':
          description: 'Error message sent, with the actual error'
          content:
            application/json:
              schema:
                type: string
                description:  no phone number searched.
        '400':
          description: 'Numero di telefono inesistente'
          content:
            application/json:
              schema:
                type: string
                description:  Numero di telefono non valido.
...
components:
  schemas:
    User:
      type: object
      required:
        - nomeCognome
        - telefono
        - email
        - password
        - checkPassword
        - token
        - logIn
        - coordX
        - coordY
      properties:
        nomeCognome:
          type: string
          description: 'Name and surname of the user'
        telefono:
          type: string
          description: 'Phone number of the user'
        email:
          type: string
          description: 'Email address of the user'
        password:
          type: string
          description: 'Password of the user'
        checkPassword:
          type: string
          description: 'Check of the users password'
        token:
          type: string
          description: 'token assosiated to the signed user'
        logIn:
          type: boolean
          description: 'Check if user is logged or not'
        coordX:
          type: string
          description: 'longitude coordinate of the user location'
        coordY:
          type: string
          description: 'latitude coordinate of the user location'
...